requirements.txt
pytest
playwright
allure-pytest
python-dotenv
pytest-rerunfailures
pytest-html